requirements.txt
pandas~=2.1.1
matplotlib==3.4.2

streamlit~=1.27.2
pillow~=10.0.1
numpy~=1.26.0
opencv-python~=4.8.1.78
keras~=2.14.0